+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * The timeline now shows object keyframes that allow editing of keyframes in that object's properties
     * Added keyboard shortcuts for editing keyframes
     * Added keyboard shortcut to insert a keyframe at the current time
+    * Menu action to repeat an animated property until the end of the composition
 * UI:
     * Timeline now allows to cycle through mask modes for layers
     * Added button to toggle keyframes in the property view